+++ include/openjij/graph/all.hpp
@@ -17,6 +17,8 @@
 #include "openjij/graph/dense.hpp"
 #include "openjij/graph/graph.hpp"
 #include "openjij/graph/polynomial.hpp"
+#include "openjij/graph/binary_polynomial_model.hpp"
+#include "openjij/graph/ising_polynomial_model.hpp"
 #include "openjij/graph/sparse.hpp"
 #include "openjij/graph/csr_sparse.hpp"
 #include "openjij/graph/square.hpp"
