+++ tests/cxxtest/graph/all.hpp
@@ -16,5 +16,6 @@
 #pragma once
 
 #include "binary_polynomial_model.hpp"
+#include "ising_polynomial_model.hpp"
 #include "quadratic.hpp"
 #include "polynomial.hpp"
